Farming out eds interface methods out of mesh_discovery.go
